Fix fake_pmap bug for functions with default parameters
This fixes a bug where using fake_pmap on a function with kwargs with
default parameters caused an error when calling the wrapped function.

Since jax.pmap allows kwargs (at least in some contexts) and jax.vmap
does not, the chex.fake_pmap context uses function inspection to
convert kwargs to args. This wrapper also supplies default arguments
as explicit args, which causes problems with jax.vmap where they are
handled as input axes. Not doing this default argument conversion
means the function is called with the arguments expected.

This also adds a new test case to cover defining and re-defining
defaulted arguments.

PiperOrigin-RevId: 342630510
Change-Id: I4bbb5bd5652ecb90503e8707e7a2dba5021f8548