Chex: asserts which only make sense on >1 tree now demand this.
Renamed functions which only make sense on multiple trees to have `_trees_` instead of `_tree_` in name. The old names still work and will be removed in future.

This removes the potential failure case of doing `assert_tree_all_close([a, b])` and that passing despite `a` and `b` being totally different.

PiperOrigin-RevId: 382693212