[JAX] Make sure that static arguments to jax.pmap are hashable in Chex.
An upcoming change to JAX means that JAX will report an error if static arguments to pmap() are not hashable. This was already a requirement for static arguments to jit(). Previously, JAX compared static arguments that were not hashable based on object identity, which is error-prone since it can easily lead to accidental cache misses for objects like lists or arrays.

PiperOrigin-RevId: 386001420