Remove `chex.{C,G,T}puDevice` in favour of `chex.Device`.
In an upcoming change the Python CpuDevice type will be removed from the XLA
extension (it is equivalent to and extends from Device).

In general we (or at least I) are trying to unify the various device
implementations such that in Python there need only be one type of Device (just
like we have `DeviceArray` and not `{C,G,T}puDeviceArray`).

PiperOrigin-RevId: 441765056