`jax.Array` is a unified type that will subsume JAX's `DeviceArray`, `ShardedDeviceArray` and `GlobalDeviceArray`.
This change replaces uses of `pxla.ShardedDeviceArray`, which is a deprecated private type, with `jax.Array` which is its public replacement.

Code that distinguishes between `ShardedDeviceArray`s and other kinds of JAX array has been changed to look for a `PmapSharding`. This is a temporary workaround that allows users to distinguish between arrays created by `pmap()` and other APIs. This method may not continue to work in the future since ultimately there may be no way to determine which API produced any given sharded `Array`.

PiperOrigin-RevId: 480125918