Fix race condition in chexify
When called with multiple threads in async mode, a chexified function and
`block_until_chexify_assertions_complete` both access the internal futures queue
without synchronization. This adds a lock around the futures queue to avoid the
circumstance where the `popleft` method can be called on an empty queue, causing
an uncaught `IndexError`.

PiperOrigin-RevId: 534012863