Add `strict` argument to `assert_trees_all_equal`.
This enables using the 'strict' mode of `numpy.testing.assert_array_equal`. This
mode (disabled by default) has stricter handling of scalar values, as described
in
https://numpy.org/doc/stable/reference/generated/numpy.testing.assert_array_equal.html. The
argument is `False` by default, so this changes allows users to enable the
stricter equality checking, if they wish.

The jittable implementation of `assert_trees_all_equal` has always been 'strict'
in the `numpy.testing` sense. I have not implemented `strict=False` for this
function, although it would probably be possible to do so.

PiperOrigin-RevId: 543676343