Add support for `kw_only` for chex.dataclass.
In practice, chex.dataclasses is keyword only, but isn't marked as such for `dataclasses`, which e.g. prevents doing inheritance across chex.dataclasses with a mix of default and not-defaulted arguments.

The feature is a no-op in Python 3.9 and lower because `kw_only` was added to `dataclasses.dataclass` only in Python 3.10.

PiperOrigin-RevId: 549313777