Restrict `assert_type` type checking when expected type is a dtype
This makes `assert_type` require an exact match when the expected type is a
dtype. The previous behaviour was for an array of dtype `bfloat16` pass
`assert_type` with an expected type of `float32`, since they are both of type
`floating`. I believe this was unintended behaviour and is confusing to users,
so we now enforce exact type matching.

PiperOrigin-RevId: 595428740