Adjust test tolerances for test_assert_trees_all_close
The test for chexify `assert_trees_all_close` asserted that `jnp.isclose(10, 20,
rtol=0.5, atol0.5)` is false. By the rules of `np.testing.assert_allclose` this
should actually pass, because (|10-20| <= 0.5 + (0.5 * 20)). However, it was only
when JAX PR #19936 was merged that this became the case.

This changes the rtol/atol to make the test pass.

PiperOrigin-RevId: 612538833