chex: avoid use of deprecated `jax.util.unzip2`
This was deprecated in JAX v0.6.0, and will be removed in JAX v0.7.0.

The main reason to use `unzip2(x)` rather than `tuple(zip(*x))` is becuase we need `tree.flatten` to return a pair even when `x` is empty. So rather than redefining `unzip2`, we catch this case explicitly.

PiperOrigin-RevId: 752901132